--- download.sh
@@ -1,4 +1,0 @@
-# download the 3D obj file
-
-wget https://dl.fbaipublicfiles.com/pytorch3d/data/dolphin/dolphin.obj
-mv dolphin.obj ./input/
