+++ pyproject.toml
@@ -1,0 +1,18 @@
+[project]
+name = "pytorch3d-  samples"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "autodistill-grounded-sam>=0.1.2",
+    "autodistill>=0.1.29",
+    "clip>=0.2.0",
+    "pip>=25.0.1",
+    "roboflow>=1.1.54",
+    "scikit-learn>=1.6.1",
+    "supervision>=0.25.1",
+    "timm>=1.0.15",
+    "transformers>=4.49.0",
+    "ultralytics>=8.3.85",
+]
