+++ src/download.sh
@@ -1,0 +1,4 @@
+# download the 3D obj file
+
+wget https://dl.fbaipublicfiles.com/pytorch3d/data/dolphin/dolphin.obj
+mv dolphin.obj ./input/
