--- test.py
@@ -1,4 +1,0 @@
-"""
-write a python class with main and commandline args
-"""
-class
